Nav Links populated
In the javascript file I've written the function that dynamically populates the navbar. It seems to be working fine, I've written step by step comments for each line of code explaining what it's for.

In the css file I've made the ul a flex container, so the nav items (after being populated by javascript) are displayed evenly across the screen horizontally.

In the html file I then added a fourth section to the document, following the pattern of the first 3, to check that everything is automatically populated into the navbar.