Show active section
Whichever section is being viewed by the user is now being highlighted (with a red background.)

In the js file I've written the highlightActive function to make it work. I've written comments to explain what each line is for.

I've given the document a scroll event listener to call the highlightActive function. This checks which section needs the active class applied every time the user scrolls.

In the css file I've give the 'your-active-class' a background colour of red, to make the highlighted class stand out.