ScrollToSection function
Here I've re-built the scrolling functionality with JavaScript.

An explanation of my implementation can be found via the comments in the index.js file.

I've also removed the bold font style for the links in the css file. This was to make more room for the links on narrow mobile screens.